requirements.txt
pandas==2.1.4
sqlalchemy==2.0.25
pymysql==1.1.0
mysqlclient==2.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apache-airflow-providers-mysql==5.5.0
apache-airflow-providers-postgres==5.10.0
pytest==7.4.0
pytest-cov==4.1.0
